requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0.0
pytest>=7.0.0
jupyter>=1.0.0
matplotlib>=3.5.0